# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/causas/{slug}": {
      "get": {
        "operationId": "CampanasController_detalle",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "CU02 · Ficha de una campaña con sus fondos",
        "tags": [
          "causas"
        ]
      }
    },
    "/api/v1/ongs/{id}": {
      "get": {
        "operationId": "CampanasController_fichaOng",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-SO-01 · Ficha de ONG con su puntaje explicado",
        "tags": [
          "causas"
        ]
      }
    },
...
    "/api/v1/ongs/{id}/fondos": {
      "get": {
        "operationId": "CampanasController_estadoFondos",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "CU12 · Estado de fondos: recaudado, retenido y ejecutado",
        "tags": [
          "causas"
        ]
      }
    },
    "/api/v1/ongs/{id}/miembros": {
      "get": {
        "operationId": "CampanasController_miembros",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Equipo de la organizacion, para su administrador",
        "tags": [
          "causas"
        ]
      },
      "post": {
        "operationId": "CampanasController_agregarMiembro",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Agregar a una persona con cuenta al equipo de la organizacion",
        "tags": [
          "causas"
        ]
      }
    },
    "/api/v1/ongs/{id}/miembros/{usuarioId}": {
      "patch": {
        "operationId": "CampanasController_cambiarMiembro",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "usuarioId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Cambiar el cargo de un miembro o desactivarlo",
        "tags": [
          "causas"
        ]
      }
    },
...
    "/api/v1/ongs/{id}/verificacion": {
      "patch": {
        "operationId": "CampanasController_verificarOng",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "CU14 · Verificar, rechazar o suspender una ONG",
        "tags": [
          "causas"
        ]
      }
    },
    "/api/v1/ongs/{id}/perceptora": {
      "patch": {
        "operationId": "CampanasController_registrarPerceptora",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-DE-07 · Registrar la calificacion de perceptora de donaciones",
        "tags": [
          "causas"
        ]
      }
    },
    "/api/v1/ongs/{id}/campanas": {
      "post": {
        "operationId": "CampanasController_crearCampana",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "CU09 · Crear una campaña",
        "tags": [
          "causas"
        ]
      }
    },
    "/api/v1/campanas/{id}": {
      "patch": {
        "operationId": "CampanasController_actualizarCampana",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-04 · Editar, publicar, pausar o cerrar una campaña",
        "tags": [
          "causas"
        ]
      }
    },
    "/api/v1/fondos/{id}": {
      "patch": {
        "operationId": "CampanasController_actualizarFondo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-05 · Editar, pausar o cerrar un fondo",
        "tags": [
          "causas"
        ]
      }
    },
    "/api/v1/campanas/{id}/fondos": {
      "post": {
        "operationId": "CampanasController_crearFondo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "CU09 · Crear un fondo con categoria de gasto y meta",
        "tags": [
          "causas"
        ]
      }
    },
...
    "/api/v1/analitica/informe/{ongId}": {
      "get": {
        "operationId": "AnaliticaController_informe",
        "parameters": [
          {
            "name": "ongId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "CU17 · Informe de auditoria de una organizacion",
        "tags": [
          "analitica"
        ]
      }
    },
    "/api/v1/analitica/exportar/libro/{fondoId}": {
      "get": {
        "operationId": "AnaliticaController_exportarLibro",
        "parameters": [
          {
            "name": "fondoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-15 · Extracto del libro de un fondo, con sus hashes",
        "tags": [
          "analitica"
        ]
      }
    },
    "/api/v1/analitica/exportar/diario/{fondoId}": {
      "get": {
        "operationId": "AnaliticaController_exportarDiario",
        "parameters": [
          {
            "name": "fondoId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-CF-05 · Libro diario de un fondo en cuentas del PCGE",
        "tags": [
          "analitica"
        ]
      }
    },
    "/api/v1/analitica/exportar/gastos/{ongId}": {
      "get": {
        "operationId": "AnaliticaController_exportarGastos",
        "parameters": [
          {
            "name": "ongId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "RF-15 · Gastos de una ONG con su verificacion",
        "tags": [
          "analitica"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ayni API",
    "description": "Micro-mecenazgo dirigido con trazabilidad total. MVP v1 sin IA: la verificacion la produce el Motor de Reglas v0, que respeta el mismo contrato de datos que AIni.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}